feat (my nfts modal): cleanup my nfts modal
