+++ src/components/ui/modals/index.ts
@@ -1,1 +1,1 @@
-export { default as FullscreenModal } from './FullscreenModal'
+export { default as MyNFTsModal } from './MyNFTsModal'
