--- src/components/ui/my-nfts-modal/index.ts
@@ -1,1 +1,0 @@
-export { default as MyNFTsModal } from './MyNFTsModal'
